+++ cloud-management/src/main/java/cn/chenjun/cloud/management/operate/impl/DestroySnapshotVolumeOperateImpl.java
@@ -6,16 +6,13 @@ import cn.chenjun.cloud.common.bean.VolumeDestroyRequest;
 import cn.chenjun.cloud.common.error.CodeException;
 import cn.chenjun.cloud.common.util.Constant;
 import cn.chenjun.cloud.common.util.ErrorCode;
-import cn.chenjun.cloud.management.annotation.Lock;
 import cn.chenjun.cloud.management.data.entity.HostEntity;
 import cn.chenjun.cloud.management.data.entity.SnapshotVolumeEntity;
 import cn.chenjun.cloud.management.data.entity.StorageEntity;
 import cn.chenjun.cloud.management.operate.bean.DestroySnapshotVolumeOperate;
-import cn.chenjun.cloud.management.util.RedisKeyUtil;
 import com.google.gson.reflect.TypeToken;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.lang.reflect.Type;
 
@@ -30,8 +27,6 @@ public class DestroySnapshotVolumeOperateImpl extends AbstractOperate<DestroySna
         super(DestroySnapshotVolumeOperate.class);
     }
 
-    @Lock(value = RedisKeyUtil.GLOBAL_LOCK_KEY, write = false)
-    @Transactional(rollbackFor = Exception.class)
     @Override
     public void operate(DestroySnapshotVolumeOperate param) {
         SnapshotVolumeEntity volume = this.snapshotVolumeMapper.selectById(param.getSnapshotVolumeId());
@@ -60,12 +55,10 @@ public class DestroySnapshotVolumeOperateImpl extends AbstractOperate<DestroySna
         }.getType();
     }
 
-    @Lock(RedisKeyUtil.GLOBAL_LOCK_KEY)
-    @Transactional(rollbackFor = Exception.class)
     @Override
     public void onFinish(DestroySnapshotVolumeOperate param, ResultUtil<Void> resultUtil) {
         SnapshotVolumeEntity volume = this.snapshotVolumeMapper.selectById(param.getSnapshotVolumeId());
-        if (volume.getStatus() == cn.chenjun.cloud.management.util.Constant.SnapshotStatus.DESTROY) {
+        if (volume != null && volume.getStatus() == cn.chenjun.cloud.management.util.Constant.SnapshotStatus.DESTROY) {
             if (resultUtil.getCode() == ErrorCode.SUCCESS) {
                 snapshotVolumeMapper.deleteById(param.getSnapshotVolumeId());
             } else {
